Fix formatting as requested by @SamVerschueren.
